extract wait_for_restart_window
